requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
Flask==3.1.1
future==1.0.0
geocoder==1.38.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
ratelim==0.1.6
requests==2.32.4
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
